+++ R/radab.R
@@ -63,8 +63,8 @@ radab <- function(ADSL, # nolint
     USUBJID = unique(ADSL$USUBJID),
     VISIT = unique(ADPC$VISIT),
     PARAM = as.factor(param_init_list$relvar1),
-    PARCAT1 = paste(unique(ADSL$ARM), "Antibody"),
-    PCTPT = "Predose",
+    PARCAT1 = "A: Drug X Antibody",
+    PCTPT = unique(ADPC$PCTPT),
     stringsAsFactors = FALSE
   )
 
@@ -85,10 +85,7 @@ radab <- function(ADSL, # nolint
         (PARAM == "Antibody titer units" & !is.na(AVAL2)) ~ AVAL2,
         TRUE ~ as.numeric(NA)
       ),
-      ISTPT = "Predose",
-      ADABLFL = "Y",
-      ADAPBLFL = "Y",
-      ABLFL = "Y"
+      ISTPT = "Predose"
     ) %>%
     dplyr::select(-c(AVAL1, AVAL2))
 
@@ -130,6 +127,18 @@ radab <- function(ADSL, # nolint
   if (length(na_vars) > 0 && na_percentage > 0 && na_percentage <= 1) {
     ADAB <- mutate_na(ds = ADAB, na_vars = na_vars, na_percentage = na_percentage) # nolint
   }
+
+  # mutate time from dose variables from ADPC to convert into Days
+  ADAB <- ADAB %>% dplyr::mutate_at(c("ARELTM1", "NRELTM1", "ARELTM2", "NRELTM2"), ~ . / 24) # nolint
+  ADAB <- ADAB %>% dplyr::mutate( # nolint
+    RELTMU = "day", # nolint
+    ADABLFL = "Y",
+    ADAPBLFL = ifelse(ACTARM == "A: Drug X" | ACTARM == "C: Combination", "Y",
+      NA
+    ),
+    ABLFL = ifelse(NRELTM1 == 0, "Y", NA)
+  )
+
 
   ADAB <- apply_metadata(ADAB, "metadata/ADAB.yml") # nolint
 }